* commit '28b1432c0258e146c2aca9f2e4b02b2678348a6d':
  Update emmc perf test tool

+++ tests/ext4/rand_emmc_perf.c
@@ -41,7 +41,7 @@ static void usage(void) {
 
 int main(int argc, char *argv[])
 {
-    long max_blocks;
+    unsigned long long max_blocks;
     int fd, fd2, write_mode = 0, iops = 0;
     struct timeval start, end, res;
     unsigned int seed;
